Extract Extension too-small tests
